pyproject.toml
[build-system]
requires = ["poetry-core>=1.6"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
name = "certbot-dns-timeweb"
version = "1.0.1"
description = "Timeweb DNS Authenticator plugin for Certbot"
authors = ["Aleksandr Ivanov <oz.sasha.ivanov@gmail.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
certbot = "^2.8"
timeweb-cloud = "^0.12.2"
tld = "^0.13"

[tool.poetry.dev-dependencies]
black = "^23.3"
isort = "^5.11"
flake8 = "^6.1"
flake8-black = "*"
flake8-pyproject = "*"

[tool.poetry.plugins."certbot.plugins"]
dns-timeweb = "certbot_dns_timeweb.authenticator:Authenticator"

[tool.black]
line-length = 120
exclude = '''
/(
  # Copied from: https://github.com/psf/black/blob/master/pyproject.toml
  | \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | venv
  | _build
  | build
  | dist
)/
'''

[tool.flake8]
# Base flake8 configuration:
# https://flake8.pycqa.org/en/latest/user/configuration.html
show-source = true
statistics = false
doctests = true

# darglint configuration:
# https://github.com/terrencepreilly/darglint
strictness = 'long'
docstring-style = 'numpy'

# Plugins:
max-complexity = 8
max-line-length = 120
max-methods = 9

# Disable some pydocstyle checks:
# Exclude some pydoctest checks globally:
ignore = 'W503'

# Due to this, Flake8 will raise E203 whitespace before ':' warnings.
# Since this warning is not PEP 8 compliant, Flake8 should be configured to ignore it
extend-ignore = 'E203'

# Excluding some directories:
exclude = [
  '.git',
  '__pycache__',
  '.venv',
  'venv',
  '.eggs',
  '*.egg',
]

# Ignoring some errors in some files:
per-file-ignores = [
  # Enable `assert` keyword and magic numbers for tests:
  'tests/*.py: S101, WPS114, WPS226, WPS432',
]

[tool.isort]
profile = 'black'
include_trailing_comma = true
use_parentheses = true
# See https://github.com/timothycrosley/isort#multi-line-output-modes
multi_line_output = 3
line_length = 120

[tool.pytest.ini_options]
# Directories that are not visited by pytest collector:
norecursedirs = [
  '.git',
  '.github',
  '.pytest_cache',
  '.tox',
  '.vscode',
  '.venv',
  'venv',
  '__pycache__',
  'hooks',
  '*.egg',
  '.eggs',
  'dist',
  'build',
  'docs',
]

# Extra options:
addopts = [
  '--strict-markers',
  '--tb=short',
  '--doctest-modules',
]

testpaths = 'tests/'

python_files = 'test_*.py'

filterwarnings = [
  'error',
  'ignore::ImportWarning:importlib',
]
